メインプロセスでゲーム本体を動かせた（描画無し）
